Initial fix for SQL functions
When creating tuples, care must be taken that the Datum values
sometimes are in logical order.  heaptuple.c is now able to create valid
tuples starting from that.  Executor is still too dumb about when to use
that -- either Tuplestore or TupleTableSlot needs to carry info about
when logical order is being used.