Fix pg_upgrade of same-version clusters
There was a thinko in the way the oldest value was being handled during
pg_upgrade; I was setting it to the same value as the next multi, but
since we need to preserve the old pg_multixact data, it actually needs
to be set to the same oldest that the old cluster has.

While fixing it, I noticed that I was complicating things uselessly by
having pg_resetxlog use the age of the oldest instead of just the value
itself ... so make it simple again.