if an xid grabs a stronger lock, don't use a multi
This reduces MultiXactId consumption a lot because EvalPlanQual grabs
FOR UPDATE locks and those are later upgraded by whatever command called
EPQ, e.g. UPDATE.  That was causing a lot of multis to be creating in
trivial load such as pgbench.

Unrelated code move around, to keep things better organized.