tighten checks for pg_upgraded shared locks
Some places were letting GetMultiXactIdMembers not fail on pg_upgraded
share-locked tuples that weren't strictly necessary; detailed examination
of the infomask lets tqual.c routines avoid calling MultiXactIdIsRunning
with such multis, which makes the code cleaner.  And in the places that
remain, we can compute permissibility before the actual call, so that we
don't have to recheck and possibly throw an error afterwards.