Let GetTupleForTrigger use a relaxed tuplock
For BEFORE UPDATE FOR EACH ROW triggers, we can let EvalPlanQual use
LockTupleNoKeyExclusive when the key columns have not been modified by
the update.  This lets the update run concurrently with key-shared
locking transactions.

The old code was using NoKeyExclusive all the time, which is bogus and
incurs into lock escalation.

Note that lock escalation is still possible, if the UPDATE command does
not modify key columns but the trigger function does.