fix serious visibility bug
We were setting HEAP_XMAX_COMMITTED on only-locked tuples after waiting
for the locking transaction to go.  But this confuses
compute_new_xmax_infomask; it's not expecting that locked tuples will
get marked xmax-committed, and it loses the HEAP_XMAX_LOCK_ONLY bit,
causing future tuple updates to mark the tuple as dead.