Return invalid PK in RelationGetIndexList too, for partitioned tables
XXX Does it work to do it only for those that have no partitions?

This is necessary so that when we DROP NOT NULL in pg_dump for its
internally generated constraints, the columns in the PK don't suddenly
become no-longer-attnotnull.  Without this, the restore of the
ALTER INDEX ATTACH for its partitions fails, because the columns in the
partitions have not been marked attnotnull.