Fix ALTER TABLE DETACH for inconsistent indexes
When a partitioned table has an index that doesn't support a constraint,
but a partition has an equivalent index that does, then a DETACH
operation would misbehave: a crash in assertion-enabled systems, and a
broken coninhcount value in production systems.

While we should reject an ATTACH of a partition with such an index, we
have failed to do so in existing releases, so adding an error in stable
releases might break existing applications.  At this point I don't even
want to reject them in master, because it'd break pg_upgrade and there
would be no easy way to fix existing databases without expensive index
rebuilds.

Also, add a test case that leaves tables in this condition, so that we
can verify that pg_upgrade continues to work if we later decide to
change the policy on the master branch.

Backpatch to all supported branches.

Reported-by: Alexander Lakhin <exclusion@gmail.com>
Reviewed-by: Tender Wang <tndrwang@gmail.com>
Discussion: https://postgr.es/m/18500-62948b6fe5522f56@postgresql.org