Fix infer_arbiter_index during concurrent index operations
Previously, we would only consider indexes marked indisvalid.  But that's
problematic during CREATE INDEX CONCURRENTLY and REINDEX CONCURRENTLY,
because we would end up with inconsistents lists of inferred indexes in
concurrent transactions, leading to spurious errors about unique key
violations (because the two transactions are operating on different indexes
for the speculative insertion tokens).

Because such indexes might not be complete, we still need uniqueness to be
verified in a different way.  We do that by requiring that at least one
index marked valid is part of the set of indexes returned.  It is that index
that is going to help ensure that the inserted tuple is indeed unique.

Fix by allowing indexes with indisvalid=false be considered as potential
arbiters in infer_arbiter_index.  Add some test cases for it.

This does not fix similar problems occurring with partitioned tables or with
named constraints.  These will be fixed in follow-up commits.

Author: Mihail Nikalayeu <mihailnikalayeu@gmail.com>
Reviewed-by: Michael Paquier <michael@paquier.xyz>
Reviewed-by: Álvaro Herrera <alvherre@kurilemu.de>
Discussion: https://postgr.es/m/CANtu0ogv+6wqRzPK241jik4U95s1pW3MCZ3rX5ZqbFdUysz7Qw@mail.gmail.com