Introduce repackdb client application.
repackdb is a client-side counterpart of the REPACK command. Like vacuumdb, it
can open multiple database connections and thus it can be beneficial if user
wants to process many tables at a time.

The application works almost identically to vacuumdb, except that it runs
REPACK instead of VACUUM. Therefore we implement it as a symbolic link to
vacuumdb and adjust vacuumdb to recognize whether it's called in "vacuum mode"
(i.e. directly) or in "repack mode" (i.e. via the link). In the latter case,
the application accepts a different set of command-line options and it
reflects the mode in log and help messages.

As for the implementation, the patch splits the main() function into
main_vacuum() and main_common() and adds a new function main_repack(). (Both
main_vacuum() and main_repack() call main_common().) Some local variables are
removed from main() first, to reduce the amount of code copied to
main_repack().

TODO

* Maybe add some more regression tests
* Test on Windows OS