Rework pg_stat_progress_repack definition
We don't need separate command identifiers for progress when we already
have the RepackCommand enum, and this makes the code much simpler.

Also, change pg_stat_progress_cluster to be a simple view on top of
pg_stat_progress_repack.  I might yet backtrack on this if I decide to
make the 'table' and 'index' columns of pg_stat_progress_cluster be of
type regclass ...